refactor: painter was renamed to views
