+++ Cargo.toml
@@ -13,7 +13,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 exclude = ["docs/*", "examples/*", "*.md"]
-include = ["**/*.rs", "Cargo.toml", "src/config/languages.config.json"]
+include = ["**/*.rs", "Cargo.toml"]
 
 [dependencies]
 clap = "3.0.5"
